+++ doc/man3/OSSL_sleep.pod
@@ -1,0 +1,37 @@
+=pod
+
+=head1 NAME
+
+OSSL_sleep - delay execution for a specified amount of milliseconds
+
+=head1 SYNOPSIS
+
+ #include <openssl/crypto.h>
+
+ void OSSL_sleep(unsigned long millis);
+
+=head1 DESCRIPTION
+
+OSSL_sleep() is a convenience function to delay execution of the calling
+thread for (at least) I<millis> milliseconds.  The delay may be lengthened
+slightly by system activity, by the time spent processing the call or by
+system timers granularity.
+
+=head1 RETURN VALUES
+
+OSSL_sleep() does not return any value.
+
+=head1 HISTORY
+
+OSSL_sleep() was added in OpenSSL 3.1.
+
+=head1 COPYRIGHT
+
+Copyright 2022 The OpenSSL Project Authors. All Rights Reserved.
+
+Licensed under the Apache License 2.0 (the "License").  You may not use
+this file except in compliance with the License.  You can obtain a copy
+in the file LICENSE in the source distribution or at
+L<https://www.openssl.org/source/license.html>.
+
+=cut
